# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DCP Platform API
  version: 5.0.0
  description: |
    DCP is Saudi Arabia's first decentralized GPU compute marketplace.
    Providers register NVIDIA GPUs, install the Python daemon, and earn SAR.
    Renters submit compute jobs (LLM inference, image generation, training) that
    run on provider hardware. DCP takes 25% fee; providers earn 75%.

    ## Currency
    All monetary amounts are in **halala** (1 SAR = 100 halala) unless the field
    name ends in `_sar`.

    ## Auth
    Three auth schemes are used depending on the caller role:

    | Role | Scheme | How |
    |------|--------|-----|
    | Provider | `ProviderApiKey` | `?key=` query param or `x-provider-key` header |
    | Renter | `RenterApiKey` | `?key=` query param or `x-renter-key` header |
    | Admin | `AdminToken` | `x-admin-token` header |

    ## Rate Limits
    | Endpoint | Limit |
    |----------|-------|
    | `POST /providers/register` | 5 / IP / hour |
    | `POST /providers/heartbeat` | 4 / IP / minute |
    | `POST /renters/register` | 5 / IP / hour |
    | `POST /jobs/submit` | 10 / renter API key / minute |
    | `POST /payments/topup` | 10 / IP / minute |
    | `GET /api/providers/marketplace` | 30 / IP / minute |
    | `GET /api/admin/*` | 100 / IP / minute |
    | All other `/api/*` | 300 / IP / minute |

  contact:
    name: DCP DevRel
    url: https://dcp.sa/docs
  x-env-vars:
    DC1_ADMIN_TOKEN:
...
      required: true
    DC1_HMAC_SECRET:
...
      required: true
...
servers:
...
components:
  securitySchemes:
    ProviderApiKey:
      type: apiKey
      in: query
      name: key
      description: Provider API key. Prefix `dcp-provider-`. Also accepted as `x-provider-key` header.
    RenterApiKey:
      type: apiKey
      in: query
      name: key
      description: Renter API key. Prefix `dcp-renter-`. Also accepted as `x-renter-key` header.
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: Admin token set via DC1_ADMIN_TOKEN env var. Also accepted as Bearer token.
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        job_id:
          type: string
          description: UUID job identifier
        job_type:
          type: string
          enum: [llm-inference, llm_inference, training, rendering, image_generation, vllm_serve, custom]
        status:
          type: string
          enum: [queued, pending, running, completed, failed, cancelled]
        submitted_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        cost_halala:
          type: integer
          description: Estimated/reserved cost at submit time
        actual_cost_halala:
          type: integer
          nullable: true
          description: Actual cost settled at completion
        provider_earned_halala:
          type: integer
          nullable: true
        dc1_fee_halala:
          type: integer
          nullable: true
...
paths:
...
  /api/providers/download:
    get:
      tags: [Providers]
      summary: Download installer script with injected API key
      security:
        - ProviderApiKey: []
      parameters:
        - in: query
          name: key
          required: true
          schema:
            type: string
        - in: query
          name: platform
          schema:
            type: string
            enum: [linux, mac, darwin, windows]
      responses:
        '200':
          description: Installer script
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
  /api/providers/download-windows-exe:
    get:
      tags: [Providers]
      summary: Download generic Windows .exe installer
      responses:
        '200':
          description: Windows installer binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Installer not yet built
...
  /api/jobs/{job_id}:
    get:
      tags: [Jobs]
      summary: Get job status and details
      security:
        - RenterApiKey: []
        - ProviderApiKey: []
        - AdminToken: []
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: string
          description: Job UUID
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '403':
          description: Access denied (not your job)
        '404':
          description: Job not found
...
  /api/admin/dashboard:
    get:
      tags: [Admin]
      summary: Platform overview stats
      security:
        - AdminToken: []
      responses:
        '200':
          description: Dashboard stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats:
                    type: object
                    properties:
                      total_providers:
                        type: integer
                      online_now:
                        type: integer
                      total_renters:
                        type: integer
                      active_renters:
                        type: integer
                      total_jobs:
                        type: integer
                      completed_jobs:
                        type: integer
                      failed_jobs:
                        type: integer
                      active_jobs:
                        type: integer
                      total_revenue_halala:
                        type: integer
                      total_dc1_fees_halala:
                        type: integer
                      today_revenue_halala:
                        type: integer
                      today_jobs:
                        type: integer
                  gpu_breakdown:
                    type: array
                    items:
                      type: object
                  recent_signups:
                    type: array
                    items:
                      type: object
                  recent_heartbeats:
                    type: array
                    items:
                      type: object
...
  /api/admin/providers/{id}:
    get:
      tags: [Admin]
      summary: Full provider detail with heartbeat metrics
      security:
        - AdminToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Provider detail with uptime, metrics, and job history
...
  /api/admin/jobs/{id}:
    get:
      tags: [Admin]
      summary: Full job detail with billing split
      security:
        - AdminToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Job ID or job_id UUID
      responses:
        '200':
          description: Job detail with provider info and billing breakdown
...